save to .mat files in batch mode
export to .mat connected in batch mode and single view mode.  Updated
STAGE struct to include each study's start date time as determined from
the .edf header (i.e. t0).  This is done in the batch run function and
also in the class_ui_marking.m file as part of the loadStages method.