corrected obtuse multichannel batch run error
When running batch mode, just prior to batch_process beginning, a check
is made for play list files which kicks off a refresh of the EDF path
check which in turn enables/disables widgets as necessary.
Unfortunately, the multiple channel button was checked to see if it is
enabled higher up in the run callback.  This would cause problems on a
second, consecutive batch mode run with non, multi sourced channel
methods.  Now it's fixed.  A value is assigned to the button and
checked instead of its 'enabled' parameter.